requirements.txt
streamlit
python-dotenv
qdrant-client
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
sentence-transformers
pypdf
